+++ CPP/Week-3/Exercise-5/ex5.cpp
@@ -2,14 +2,14 @@
 
 #include <iostream>
 using namespace std;
-bool Palindrom(long int num)
+bool Palindrom(int num)
 {
-  long int numarOriginal = num;
-  long int numarInversat = 0;
+  int numarOriginal = num;
+  int numarInversat = 0;
 
   while (num > 0)
   {
-    long int cifra = num % 10;
+    int cifra = num % 10;
     numarInversat = numarInversat * 10 + cifra;
     num /= 10;
   }
@@ -18,12 +18,12 @@ bool Palindrom(long int num)
 
 int main()
 {
-  long int a, b;
+  int a, b;
   cout << "Introduceti valoare minima pentru intervalul de numere: ";
   cin >> a;
   cout << "Introduceti valoarea maxima pentru intervalul de numere: ";
   cin >> b;
-  for (long int num = a; num <= b; num++)
+  for (int num = a; num <= b; num++)
   {
     if (Palindrom(num))
     {
